+++ src/librpbase/byteswap_mmx.c
@@ -19,7 +19,7 @@
  * @param ptr Pointer to array to swap. (MUST be 16-bit aligned!)
  * @param n Number of bytes to swap. (Must be divisible by 2; an extra odd byte will be ignored.)
  */
-void __byte_swap_16_array_mmx(uint16_t *ptr, unsigned int n)
+void __byte_swap_16_array_mmx(uint16_t *ptr, size_t n)
 {
 	// Verify the block is 16-bit aligned
 	// and is a multiple of 2 bytes.
@@ -76,7 +76,7 @@ void __byte_swap_16_array_mmx(uint16_t *ptr, unsigned int n)
  * @param ptr Pointer to array to swap. (MUST be 32-bit aligned!)
  * @param n Number of bytes to swap. (Must be divisible by 4; extra bytes will be ignored.)
  */
-void __byte_swap_32_array_mmx(uint32_t *ptr, unsigned int n)
+void __byte_swap_32_array_mmx(uint32_t *ptr, size_t n)
 {
 	// Verify the block is 32-bit aligned
 	// and is a multiple of 4 bytes.
